+++ InscryptionAPI/InscryptionAPIPlugin.cs
@@ -30,7 +30,7 @@ public class InscryptionAPIPlugin : BaseUnityPlugin
 {
     public const string ModGUID = "cyantist.inscryption.api";
     public const string ModName = "InscryptionAPI";
-    public const string ModVer = "2.21.1";
+    public const string ModVer = "2.22.0";
 
     public static string Directory = "";