+++ InscryptionAPI/TalkingCards/CustomPaperTalkingCard.cs
@@ -26,3 +26,20 @@ public abstract class CustomPaperTalkingCard : PaperTalkingCard, ITalkingCard
         yield break;
     }
 }
+
+public abstract class CustomDiskTalkingCard : DiskTalkingCard, ITalkingCard
+{
+    public override DialogueEvent.Speaker SpeakerType => DialogueEvent.Speaker.Stoat;
+    public abstract string CardName { get; }
+    public abstract List<EmotionData> Emotions { get; }
+    public abstract FaceInfo FaceInfo { get; }
+    public abstract SpecialTriggeredAbility DialogueAbility { get; }
+
+    public override string OnDrawnFallbackDialogueId => OnDrawnDialogueId;
+    public override IEnumerator OnShownForCardSelect(bool forPositiveEffect)
+    {
+        yield return new WaitForEndOfFrame();
+        yield return base.OnShownForCardSelect(forPositiveEffect);
+        yield break;
+    }
+}